Override 'touch-action' for 'vertical-scroll'
With "vertical-scroll 'none'" as a feature set for an <iframe>, we want
to make sure the <iframe> cannot prevent the user from vertically
scrolling the page. One way of blocking user is to exclude "pan-y" from
"touch-action".

This CL makes sure that when an element is inside such frames, the
"touch-action" always includes "pan-y".

Link to explainer/design document:
https://docs.google.com/document/d/1qiWelnMlsOHuT_CQ0Zm_qEAf54HS5DhoIvEDHBlfqps/edit?usp=sharing

Bug: 611982
Change-Id: Ie60c665de3c82ec2438bcea10f56b54d51084eda